refactor: JS → JSX 파일 변환 (CRA->Vite)
