+++ src/features/auth/pages/KakaoCallback.jsx
@@ -5,9 +5,9 @@ import Loading from "../../../components/Loading";
 
 const KakaoCallback = () => {
   const BASE_URL =
-    process.env.NODE_ENV === "production"
-      ? process.env.REACT_APP_WWWM_BE_ENDPOINT
-      : process.env.REACT_APP_WWWM_BE_DEV_EP;
+    import.meta.env.PROD
+      ? import.meta.env.VITE_WWWM_BE_ENDPOINT
+      : import.meta.env.VITE_WWWM_BE_DEV_EP;
 
   const navigate = useNavigate();
   const location = useLocation();
