+++ src/features/calendar/pages/GetAppointmentRedirect/index.jsx
@@ -86,9 +86,9 @@ const GetAppointmentRedirect = () => {
       }
 
       const BASE_URL =
-        process.env.NODE_ENV === "production"
-          ? process.env.REACT_APP_WWWM_BE_ENDPOINT
-          : process.env.REACT_APP_WWWM_BE_DEV_EP;
+        import.meta.env.PROD
+          ? import.meta.env.VITE_WWWM_BE_ENDPOINT
+          : import.meta.env.VITE_WWWM_BE_DEV_EP;
 
       // 토큰 체크 및 인증 처리
       const token = localStorage.getItem("authToken");
