+++ src/features/user/pages/EditProfile.jsx
@@ -152,9 +152,9 @@ const EditProfile = () => {
 
     try {
       const BASE_URL =
-        process.env.NODE_ENV === "production"
-          ? process.env.REACT_APP_WWWM_BE_ENDPOINT
-          : process.env.REACT_APP_WWWM_BE_DEV_EP;
+        import.meta.env.PROD
+          ? import.meta.env.VITE_WWWM_BE_ENDPOINT
+          : import.meta.env.VITE_WWWM_BE_DEV_EP;
 
       const token = localStorage.getItem("authToken");
